requirements.txt
numpy
scipy
sounddevice
requests
beautifulsoup4
pillow
python-dotenv
PyJWT
google-genai
google-generativeai
PySide6
playwright
pyautogui
pyperclip
psutil
pynput
duckduckgo-search
yt-dlp
faster-whisper
openai-whisper
torch
pytest
keyring
opencv-python
mss
pycaw
win10toast          
send2trash 
youtube-transcript-api
pywinauto
pyaudio